refactor: run integration tests and examples against production for table level IAM Policy
