+++ google/cloud/bigtable/examples/table_admin_async_snippets.cc
@@ -294,134 +294,6 @@ void AsyncWaitForConsistency(google::cloud::bigtable::TableAdmin admin,
   (std::move(admin), std::move(cq), argv.at(0), argv.at(1));
 }
 
-void AsyncGetIamPolicy(google::cloud::bigtable::TableAdmin admin,
-                       google::cloud::bigtable::CompletionQueue cq,
-                       std::vector<std::string> argv) {
-  //! [async get iam policy]
-  namespace cbt = google::cloud::bigtable;
-  using google::cloud::future;
-  using google::cloud::StatusOr;
-  using google::iam::v1::Policy;
-  [](cbt::TableAdmin admin, cbt::CompletionQueue cq, std::string table_id) {
-    future<StatusOr<google::iam::v1::Policy>> policy_future =
-        admin.AsyncGetIamPolicy(cq, table_id);
-
-    future<void> final =
-        policy_future.then([](future<StatusOr<google::iam::v1::Policy>> f) {
-          StatusOr<google::iam::v1::Policy> iam = f.get();
-          if (!iam) throw std::runtime_error(iam.status().message());
-          std::cout << "IamPolicy details : " << iam->DebugString() << "\n";
-        });
-
-    final.get();  // block to keep the example simple
-  }
-  //! [async get iam policy]
-  (std::move(admin), std::move(cq), argv.at(0));
-}
-
-void AsyncSetIamPolicy(google::cloud::bigtable::TableAdmin admin,
-                       google::cloud::bigtable::CompletionQueue cq,
-                       std::vector<std::string> argv) {
-  //! [async set iam policy]
-  namespace cbt = google::cloud::bigtable;
-  using google::cloud::future;
-  using google::cloud::StatusOr;
-  using google::iam::v1::Policy;
-  [](cbt::TableAdmin admin, cbt::CompletionQueue cq, std::string table_id,
-     std::string role, std::string member) {
-    future<StatusOr<google::iam::v1::Policy>> updated_future =
-        admin.AsyncGetIamPolicy(cq, table_id)
-            .then([cq, admin, role, member,
-                   table_id](future<StatusOr<google::iam::v1::Policy>>
-                                 current_future) mutable {
-              auto current = current_future.get();
-              if (!current) {
-                return google::cloud::make_ready_future<
-                    StatusOr<google::iam::v1::Policy>>(current.status());
-              }
-              // This example adds the member to all existing bindings for
-              // that role. If there are no such bindgs, it adds a new one.
-              // This might not be what the user wants, e.g. in case of
-              // conditional bindings.
-              size_t num_added = 0;
-              for (auto& binding : *current->mutable_bindings()) {
-                if (binding.role() == role) {
-                  binding.add_members(member);
-                  ++num_added;
-                }
-              }
-              if (num_added == 0) {
-                *current->add_bindings() = cbt::IamBinding(role, {member});
-              }
-              return admin.AsyncSetIamPolicy(cq, table_id, *current);
-            });
-    // Show how to perform additional work while the long running operation
-    // completes. The application could use future.then() instead.
-    std::cout << "Waiting for IAM policy update to complete " << std::flush;
-    updated_future.wait_for(std::chrono::seconds(2));
-    auto result = updated_future.get();
-    std::cout << '.' << std::flush;
-    if (!result) throw std::runtime_error(result.status().message());
-    using cbt::operator<<;
-    std::cout << "DONE, the IAM Policy for " << table_id << " is\n"
-              << *result << "\n";
-  }
-  //! [async set iam policy]
-  (std::move(admin), std::move(cq), argv.at(0), argv.at(1), argv.at(2));
-}
-
-void AsyncTestIamPermissions(google::cloud::bigtable::TableAdmin admin,
-                             google::cloud::bigtable::CompletionQueue cq,
-                             std::vector<std::string> argv) {
-  //! [async test iam permissions]
-  namespace cbt = google::cloud::bigtable;
-  using google::cloud::future;
-  using google::cloud::StatusOr;
-  [](cbt::TableAdmin admin, cbt::CompletionQueue cq, std::string resource,
-     std::vector<std::string> permissions) {
-    future<StatusOr<std::vector<std::string>>> permissions_future =
-        admin.AsyncTestIamPermissions(cq, resource, permissions);
-    // Show how to perform additional work while the long running operation
-    // completes. The application could use permissions_future.then() instead.
-    std::cout << "Waiting for TestIamPermissions " << std::flush;
-    permissions_future.wait_for(std::chrono::seconds(2));
-    std::cout << '.' << std::flush;
-    auto result = permissions_future.get();
-    if (!result) throw std::runtime_error(result.status().message());
-    std::cout << "DONE, the current user has the following permissions [";
-    char const* sep = "";
-    for (auto const& p : *result) {
-      std::cout << sep << p;
-      sep = ", ";
-    }
-    std::cout << "]\n";
-  }
-  //! [async test iam permissions]
-  (std::move(admin), std::move(cq), argv.at(0), {argv.begin() + 1, argv.end()});
-}
-
-void AsyncTestIamPermissionsCommand(std::vector<std::string> argv) {
-  if (argv.size() < 3) {
-    throw Usage{
-        "async-test-iam-permissions <project-id> <instance-id> <resource-id>"
-        "<permission> [permission ...]"};
-  }
-  auto const project_id = argv[0];
-  auto const instance_id = argv[1];
-  argv.erase(argv.begin(), argv.begin() + 2);
-
-  google::cloud::CompletionQueue cq;
-  std::thread th([&cq] { cq.Run(); });
-  google::cloud::bigtable::examples::AutoShutdownCQ shutdown(cq, std::move(th));
-
-  google::cloud::bigtable::TableAdmin admin(
-      google::cloud::bigtable::CreateDefaultAdminClient(
-          project_id, google::cloud::bigtable::ClientOptions{}),
-      instance_id);
-
-  AsyncTestIamPermissions(admin, cq, std::move(argv));
-}
-
 void RunAll(std::vector<std::string> const& argv) {
   namespace examples = ::google::cloud::bigtable::examples;
   namespace cbt = google::cloud::bigtable;
@@ -431,17 +303,12 @@ void RunAll(std::vector<std::string> const& argv) {
   examples::CheckEnvironmentVariablesAreSet({
       "GOOGLE_CLOUD_PROJECT",
       "GOOGLE_CLOUD_CPP_BIGTABLE_TEST_INSTANCE_ID",
-      "GOOGLE_CLOUD_CPP_BIGTABLE_TEST_SERVICE_ACCOUNT",
   });
   auto const project_id =
       google::cloud::internal::GetEnv("GOOGLE_CLOUD_PROJECT").value();
   auto const instance_id = google::cloud::internal::GetEnv(
                                "GOOGLE_CLOUD_CPP_BIGTABLE_TEST_INSTANCE_ID")
                                .value();
-  auto const service_account =
-      google::cloud::internal::GetEnv(
-          "GOOGLE_CLOUD_CPP_BIGTABLE_TEST_SERVICE_ACCOUNT")
-          .value();
 
   cbt::TableAdmin admin(
       cbt::CreateDefaultAdminClient(project_id, cbt::ClientOptions{}),
@@ -494,24 +361,6 @@ void RunAll(std::vector<std::string> const& argv) {
 
   std::cout << "\nRunning the AsyncDropAllRows() example" << std::endl;
   AsyncDropAllRows(admin, cq, {table_id});
-
-  // TODO(#151) - remove workarounds for emulator bug(s).
-  if (!UsingEmulator()) {
-    std::cout << "\nRunning AsyncGetIamPolicy() example" << std::endl;
-    AsyncGetIamPolicy(admin, cq, {table_id});
-
-    std::cout << "\nRunning AsyncSetIamPolicy() example" << std::endl;
-    AsyncSetIamPolicy(
-        admin, cq,
-        {table_id, "roles/bigtable.user", "serviceAccount:" + service_account});
-
-    std::cout << "\nRunning AsyncTestIamPermissions() example [1]" << std::endl;
-    AsyncTestIamPermissionsCommand(
-        {project_id, instance_id, table_id, "bigtable.tables.get"});
-
-    std::cout << "\nRunning AsyncTestIamPermissions() example [2]" << std::endl;
-    AsyncTestIamPermissions(admin, cq, {table_id, "bigtable.tables.get"});
-  }
 
   std::cout << "\nRunning the AsyncDeleteTable() example" << std::endl;
   AsyncDeleteTable(admin, cq, {table_id});
