+++ google/cloud/bigtable/tests/admin_integration_test.cc
@@ -30,7 +30,6 @@ namespace bigtable = google::cloud::bigtable;
 class AdminIntegrationTest : public bigtable::testing::TableIntegrationTest {
  protected:
   std::unique_ptr<bigtable::TableAdmin> table_admin_;
-  std::string service_account_;
 
   void SetUp() {
     if (google::cloud::internal::GetEnv(
@@ -39,10 +38,6 @@ class AdminIntegrationTest : public bigtable::testing::TableIntegrationTest {
       GTEST_SKIP();
     }
     TableIntegrationTest::SetUp();
-    service_account_ = google::cloud::internal::GetEnv(
-                           "GOOGLE_CLOUD_CPP_BIGTABLE_TEST_SERVICE_ACCOUNT")
-                           .value_or("");
-    ASSERT_FALSE(service_account_.empty());
 
     std::shared_ptr<bigtable::AdminClient> admin_client =
         bigtable::CreateDefaultAdminClient(
@@ -316,51 +311,6 @@ TEST_F(AdminIntegrationTest, WaitForConsistencyCheck) {
   // Cleanup the table and the instance.
   EXPECT_STATUS_OK(table_admin.DeleteTable(random_table_id));
   EXPECT_STATUS_OK(instance_admin.DeleteInstance(id));
-}
-
-/// @test Verify that IAM Policy APIs work as expected.
-TEST_F(AdminIntegrationTest, SetGetTestIamAPIsTest) {
-  // TODO(#151) - remove workarounds for emulator bugs(s)
-  if (UsingCloudBigtableEmulator()) GTEST_SKIP();
-
-  using GC = bigtable::GcRule;
-  std::string const table_id = RandomTableId();
-
-  // verify new table id in current table list
-  auto previous_table_list =
-      table_admin_->ListTables(btadmin::Table::NAME_ONLY);
-  ASSERT_STATUS_OK(previous_table_list);
-  auto previous_count = CountMatchingTables(table_id, *previous_table_list);
-  ASSERT_EQ(0, previous_count) << "Table (" << table_id << ") already exists."
-                               << " This is unexpected, as the table ids are"
-                               << " generated at random.";
-  // create table config
-  bigtable::TableConfig table_config(
-      {{"fam", GC::MaxNumVersions(5)},
-       {"foo", GC::MaxAge(std::chrono::hours(24))}},
-      {"a1000", "a2000", "b3000", "m5000"});
-
-  // create table
-  ASSERT_STATUS_OK(table_admin_->CreateTable(table_id, table_config));
-
-  auto iam_policy = bigtable::IamPolicy({bigtable::IamBinding(
-      "roles/bigtable.reader", {"serviceAccount:" + service_account_})});
-
-  auto initial_policy = table_admin_->SetIamPolicy(table_id, iam_policy);
-  ASSERT_STATUS_OK(initial_policy);
-
-  auto fetched_policy = table_admin_->GetIamPolicy(table_id);
-  ASSERT_STATUS_OK(fetched_policy);
-
-  EXPECT_EQ(initial_policy->version(), fetched_policy->version());
-  EXPECT_EQ(initial_policy->etag(), fetched_policy->etag());
-
-  auto permission_set = table_admin_->TestIamPermissions(
-      table_id, {"bigtable.tables.get", "bigtable.tables.readRows"});
-  ASSERT_STATUS_OK(permission_set);
-
-  EXPECT_EQ(2, permission_set->size());
-  EXPECT_STATUS_OK(table_admin_->DeleteTable(table_id));
 }
 // Test Cases Finished
